requirements.txt
PyTomCrypt==0.9.0
simplejson==3.3.1
cython
-e git+https://github.com/surfly/gevent.git#egg=gevent
